Bit of Refactoring
Added MVVM for Channels Support
TODO: actually hook up to a file!
