+++ DCS-SR-Client/Audio/Providers/AudioProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Ciribob.DCS.SimpleRadio.Standalone.Client.Settings;
 using Ciribob.DCS.SimpleRadio.Standalone.Client.UI;
 using Ciribob.DCS.SimpleRadio.Standalone.Common;
 
@@ -6,11 +7,11 @@ namespace Ciribob.DCS.SimpleRadio.Standalone.Client.Audio
 {
     public abstract class AudioProvider
     {
-        protected readonly Settings _settings;
+        protected readonly Settings.SettingsStore _settings;
 
         public AudioProvider()
         {
-            _settings = Settings.Instance;
+            _settings = Settings.SettingsStore.Instance;
         }
