+++ DCS-SR-Client/Audio/Providers/JitterBufferProviderInterface.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Ciribob.DCS.SimpleRadio.Standalone.Client.Audio.Managers;
 using NAudio.Utils;
 using NAudio.Wave;
