+++ DCS-SR-Client/Settings/RadioChannels/PresetChannel.cs
@@ -1,0 +1,14 @@
+﻿namespace Ciribob.DCS.SimpleRadio.Standalone.Client.UI.ClientWindow.PresetChannels
+{
+    public class PresetChannel
+    {
+
+        public string Text { get; set; }
+        public object Value { get; set; }
+
+        public override string ToString()
+        {
+            return Text;
+        }
+    }
+}
