--- DCS-SR-Client/UI/ClientWindow/FavouriteServersView.xaml
@@ -1,111 +1,0 @@
-﻿<UserControl x:Class="Ciribob.DCS.SimpleRadio.Standalone.Client.UI.ClientWindow.FavouriteServersView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:clientWindow="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Client.UI.ClientWindow"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:ui="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Client.UI"
-             d:DataContext="{d:DesignInstance clientWindow:FavouriteServersViewModel,
-                                              IsDesignTimeCreatable=False}"
-             d:DesignHeight="300"
-             d:DesignWidth="300"
-             mc:Ignorable="d">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition />
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
-
-        <TextBlock Grid.Row="0"
-                   Grid.Column="0"
-                   Margin="0,0,0,5"
-                   HorizontalAlignment="Center"
-                   Text="Add new configuration" />
-
-        <Grid Grid.Row="1">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition />
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition />
-                <ColumnDefinition Width="Auto" />
-            </Grid.ColumnDefinitions>
-
-            <TextBlock Grid.Column="1"
-                       Margin="0,0,5,0"
-                       VerticalAlignment="Center"
-                       Text="Name" />
-            <TextBox Grid.Column="2" Text="{Binding Path=NewName}" />
-            <TextBlock Grid.Column="3"
-                       Margin="10,0,5,0"
-                       VerticalAlignment="Center"
-                       Text="ServerAddress" />
-            <TextBox Grid.Column="4" Text="{Binding Path=NewAddress}" />
-            <Button Grid.Column="5"
-                    Height="23"
-                    Margin="5,0,0,0"
-                    Command="{Binding Path=NewAddressCommand}">
-                <Path Data="F1 M 2.88,17.2725C 0.959991,15.3475 0,12.95 0,10.08C 0,7.21002 0.959991,4.8125 2.88,2.88751C 4.80002,0.962494 7.19125,0 10.0537,0C 12.9163,0 15.3163,0.962494 17.2538,2.88751C 19.1913,4.8125 20.16,7.21002 20.16,10.08C 20.16,12.95 19.1913,15.3475 17.2538,17.2725C 15.3163,19.1975 12.9163,20.16 10.0537,20.16C 7.19125,20.16 4.80002,19.1975 2.88,17.2725 Z M 3.84,11.3625L 8.81998,11.3625L 8.81998,16.32L 11.34,16.32L 11.34,11.3625L 16.32,11.3625L 16.32,8.85001L 11.34,8.85001L 11.34,3.84L 8.81998,3.84L 8.81998,8.85001L 3.84,8.85001L 3.84,11.3625 Z "
-                      Fill="Black"
-                      Stretch="Uniform"
-                      UseLayoutRounding="False" />
-            </Button>
-        </Grid>
-
-        <Button Grid.Row="2"
-                Width="120"
-                Margin="0,30,0,10"
-                Command="{Binding Path=RemoveSelectedCommand}"
-                Content="Remove Selected" />
-
-        <DataGrid Name="FavouritesGrid"
-                  Grid.Row="3"
-                  HorizontalAlignment="Center"
-                  AutoGenerateColumns="False"
-                  ItemsSource="{Binding Path=Addresses}"
-                  SelectedItem="{Binding Path=SelectedItem}">
-            <DataGrid.Columns>
-                <DataGridTextColumn Binding="{Binding Path=Name}" Header="Name" />
-                <DataGridTextColumn Binding="{Binding Path=Address}" Header="Ip ServerAddress (and port)" />
-                <DataGridTemplateColumn Header="Is Default">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="ui:ServerAddress">
-
-                            <Button Name="Button"
-                                    Command="{Binding ElementName=FavouritesGrid,
-                                                      Path=DataContext.OnDefaultChangedCommand}"
-                                    CommandParameter="{Binding}"
-                                    Padding="0" />
-
-                            <DataTemplate.Triggers>
-                                <DataTrigger Binding="{Binding Path=IsDefault}" Value="True">
-                                    <DataTrigger.Setters>
-                                        <Setter TargetName="Button" Property="Content">
-                                            <Setter.Value>
-                                                <Path Height="15"
-                                                      Data="F1 M 5.33334,3.05176e-005C 8.27869,3.05176e-005 10.6667,2.38669 10.6667,5.33334L 10.6667,128C 10.6667,130.947 8.27869,133.333 5.33334,133.333C 2.38669,133.333 0,130.947 0,128L 0,5.33334C 0,2.38669 2.38669,3.05176e-005 5.33334,3.05176e-005 Z M 58.4986,3.84378C 65.2643,6.4642 70.6588,8.52805 80,4.45712L 80,44.6909C 70.2527,48.5827 64.0612,46.1921 57.5364,43.664C 52.8828,41.8627 48.0729,40 41.3333,40C 27.6667,40 17.9681,48.1263 16,49.9122L 16,9.08273C 20.028,5.89725 29.1282,3.05176e-005 41.3333,3.05176e-005C 48.5703,3.05176e-005 53.845,2.04163 58.4986,3.84378 Z "
-                                                      Fill="Black"
-                                                      Stretch="Uniform" />
-                                            </Setter.Value>
-                                        </Setter>
-                                    </DataTrigger.Setters>
-                                </DataTrigger>
-                            </DataTemplate.Triggers>
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-            </DataGrid.Columns>
-        </DataGrid>
-
-        <Button Grid.Row="4"
-                Height="40"
-                Margin="0,5,0,0"
-                Command="{Binding Path=SaveCommand}"
-                Content="Save" />
-    </Grid>
-
-</UserControl>
