+++ DCS-SR-Client/UI/RadioOverlayWindow/PresetChannels/PresetChannelsView.xaml
@@ -1,0 +1,45 @@
+﻿<UserControl x:Class="Ciribob.DCS.SimpleRadio.Standalone.Client.UI.RadioOverlayWindow.PresetChannels.PresetChannelsView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Client.UI.ClientWindow.PresetChannels"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:presetChannels="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Client.UI.RadioOverlayWindow.PresetChannels"
+             d:DataContext="{d:DesignInstance presetChannels:PresetChannelsViewModel,
+                                              IsDesignTimeCreatable=False}"
+             d:DesignHeight="100"
+             d:DesignWidth="135"
+             mc:Ignorable="d">
+    <WrapPanel HorizontalAlignment="Center"
+               Background="#444"
+               Orientation="Vertical">
+        <ComboBox x:Name="Mic"
+                  Width="135"
+                  HorizontalAlignment="Center"
+                  VerticalAlignment="Top"
+                  ItemsSource="{Binding Path=PresetChannels}"
+                  SelectedItem="{Binding Path=SelectedPresetChannel}" />
+        <WrapPanel Width="135"
+                   HorizontalAlignment="Center"
+                   VerticalAlignment="Center"
+                   Background="#444"
+                   Orientation="Horizontal">
+            <Button Name="ReloadButton"
+                    Width="50"
+                    Height="20"
+                    Margin="15,5,0,0"
+                    HorizontalAlignment="Center"
+                    Command="{Binding Path=ReloadCommand}"
+                    Content="Reload"
+                    Style="{x:Null}" />
+            <Button Name="LoadFromFileButton"
+                    Width="50"
+                    Height="20"
+                    Margin="5,5,0,0"
+                    HorizontalAlignment="Center"
+                    Command="{Binding Path=LoadFromFileCommand}"
+                    Content="Load File"
+                    Style="{x:Null}" />
+        </WrapPanel>
+    </WrapPanel>
+</UserControl>
