+++ DCS-SR-Client/UI/RadioOverlayWindow/PresetChannels/PresetChannelsView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Ciribob.DCS.SimpleRadio.Standalone.Client.UI.RadioOverlayWindow.PresetChannels
+{
+    /// <summary>
+    /// Interaction logic for PresetChannelsView.xaml
+    /// </summary>
+    public partial class PresetChannelsView : UserControl
+    {
+        public PresetChannelsView()
+        {
+            InitializeComponent();
+        }
+    }
+}
