+++ DCS-SR-Common/DCSState/RadioInformation.cs
@@ -52,6 +52,8 @@
 
         public bool expansion = false;
 
+        public int channel = -1;
+
         /**
          * Used to determine if we should send an update to the server or not
          * We only need to do that if something that would stop us Receiving happens which
